+++ core/src/com/angian/libgdx/pipes/WaterTimer.java
@@ -8,22 +8,33 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 public class WaterTimer extends BaseActor {
     private float origTime;
     private float remainingTime;
+    private boolean started = false;
     private boolean expired = false;
     private float origHeight;
 
     private static ShapeRenderer shapeRenderer = new ShapeRenderer();
 
 
-    public WaterTimer(float time2water, Rectangle layout, Stage s) {
+    public WaterTimer(Rectangle layout, Stage s) {
         super(layout.x, layout.y, s);
         setWidth(layout.width);
         setHeight(layout.height);
 
-        remainingTime = time2water;
-        origTime = time2water;
         origHeight = layout.height;
+
+        origTime = Float.MAX_VALUE;
+        remainingTime = Float.MAX_VALUE;
     }
 
+    public void start(float time2water) {
+        origTime = time2water;
+        remainingTime = time2water;
+        started = true;
+    }
+
+    public boolean isStarted() {
+        return started;
+    }
 
     public boolean isExpired() {
         return expired;
@@ -32,7 +43,7 @@ public class WaterTimer extends BaseActor {
     @Override
     public void act(float dt) {
         remainingTime -= dt;
-        if (remainingTime <= 0) {
+        if (started && remainingTime <= 0) {
             remainingTime = 0;
             expired = true;
         }